ch: Sync device addresses between persistent and runtime config
When hotplugging device, they might end up with different address. For
example, up to now first htoplugging a device non-persistently and the
hotplugging a device persistently would lead to different BDF
assignment. This is because the non-persistent device is only added to
the runtime config and receives the next free BDF. Hotplugging the next
device but now persistently would do the same, but htis time in both
configurations, which would lead to the case that the BDF of ther
persistent configuration would be one off from the state in the runtime
config. We synchronize device addresses of the persistent config
therefore with those of the runtime config.

Signed-off-by: Pascal Scholz <pascal.scholz@cyberus-technology.de>
On-behalf-of: SAP pascal.scholz@sap.com