ch: Ensure live and persistent stay in sync when detaching devices
Right now the live and persistent config can go out of sync wrt. device
addresses when devices are detached. For example, adding a device to
both, then only removing it in the live config leaves the address
allocated in the persistent config, while it will be marked free in the
live config. The next device attached will then receive different BDF in
the live and persistent config. We fix this by freeing the address only
if the device it self is transient and thus was never added to the
transient config, or if both configs are affected.

Signed-off-by: Pascal Scholz <pascal.scholz@cyberus-technology.de>
On-behalf-of: SAP pascal.scholz@sap.com